Added test for overviewItem creation
Also changed the final monetary sum to BigDecimal to allow rounding.

+++ MoneyManagementBusiness/src/main/kotlin/com/mauquoi/money/model/OverviewItem.kt
@@ -1,8 +1,9 @@
 package com.mauquoi.money.model
 
+import java.math.BigDecimal
 import java.util.*
 
-data class OverviewItem(val mainCurrencyValue: Double,
+data class OverviewItem(val mainCurrencyValue: BigDecimal,
                         val mainCurrency: Currency,
-                        val distribution: Map<Currency, Double>
+                        val distribution: Map<Currency, BigDecimal>
 )
